rename: [step4] AddRestaurantModal 네이밍 변경
